+++ test/onnx/test_pytorch_onnx_caffe2.py
@@ -336,7 +336,7 @@ class TestCaffe2Backend_opset9(unittest.TestCase):
 
         # test that the model still runs with a different batch size
         # (save the model with a batch_size of 1 with rnn with a variable batch size,
-        # othewise expand will fail)
+        # othewise expand will fail) 
         variable_batch_size_init_input = make_input(1)
         onnxir, _ = do_export(model, variable_batch_size_init_input, keep_initializers_as_inputs=True)
         other_input = make_input(RNN_BATCH_SIZE + 1)
@@ -379,7 +379,7 @@ class TestCaffe2Backend_opset9(unittest.TestCase):
 
         # test that the model still runs with a different batch size
         # (save the model with a batch_size of 1 with rnn with a variable batch size,
-        # othewise expand will fail)
+        # othewise expand will fail) 
         variable_batch_size_init_input = make_input(1)
         onnxir, _ = do_export(model, variable_batch_size_init_input, keep_initializers_as_inputs=True)
         other_input = make_input(RNN_BATCH_SIZE + 1)
@@ -420,7 +420,7 @@ class TestCaffe2Backend_opset9(unittest.TestCase):
 
         # test that the model still runs with a different batch size
         # (save the model with a batch_size of 1 with rnn with a variable batch size,
-        # othewise expand will fail)
+        # othewise expand will fail) 
         variable_batch_size_init_input = make_input(1)
         onnxir, _ = do_export(model, variable_batch_size_init_input, keep_initializers_as_inputs=True)
         other_input = make_input(RNN_BATCH_SIZE + 1)
@@ -1489,7 +1489,7 @@ class TestCaffe2Backend_opset9(unittest.TestCase):
         x = torch.randn(16, 3, 256, 256)
         self.run_model_test(ReduceSumMultipleAxes(), train=False, input=(x,), batch_size=BATCH_SIZE, use_gpu=False)
 
-    # InstanceNorm model (used in the subgraph) includes unused weights,
+    # InstanceNorm model (used in the subgraph) includes unused weights, 
     # so skip this in TestCaffe2BackendEmbed
     @skipIfEmbed
     def test_group_norm(self):
@@ -2436,90 +2436,6 @@ TestCaffe2BackendEmbed_opset10 = type(str("TestCaffe2BackendEmbed_opset10"),
                                       (unittest.TestCase,),
                                       dict(TestCaffe2Backend_opset9.__dict__,
                                            embed_params=True, opset_version=10))
-
-
-class TestQuantizedOps(unittest.TestCase):
-    def generic_test(self, model, sample_inputs, input_names=None):
-        pt_inputs = tuple(torch.from_numpy(x) for x in sample_inputs)
-        model.qconfig = torch.quantization.default_qconfig
-        q_model = torch.quantization.prepare(model, inplace=False)
-        q_model = torch.quantization.convert(q_model, inplace=False)
-        pytorch_res = q_model(*pt_inputs)
-        f = io.BytesIO()
-        torch.onnx.export(q_model, pt_inputs, f, input_names=input_names, operator_export_type=torch.onnx.OperatorExportTypes.ONNX_ATEN_FALLBACK)
-        f.seek(0)
-        onnx_model = onnx.load(f)
-        caffe_res = c2.run_model(onnx_model, dict(zip(input_names, sample_inputs)))[0]
-        np.testing.assert_almost_equal(pytorch_res.numpy(), caffe_res, decimal=3)
-
-    def generic_unary_test(self, op):
-        class QModule(torch.nn.Module):
-            def __init__(self, op):
-                super(QModule, self).__init__()
-                self.quant1 = torch.quantization.QuantStub()
-                self.op = op
-                self.dequant = torch.quantization.DeQuantStub()
-
-            def forward(self, x):
-                res = self.op(self.quant1(x))
-                return self.dequant(res)
-
-        x = np.random.random((1, 2)).astype("float32")
-        self.generic_test(QModule(op), (x,), input_names=["x"])
-
-    def test_quantized_add(self):
-        class QAddModule(torch.nn.Module):
-            def __init__(self):
-                super(QAddModule, self).__init__()
-                self.quant1 = torch.quantization.QuantStub()
-                self.quant2 = torch.quantization.QuantStub()
-                self.dequant = torch.quantization.DeQuantStub()
-
-            def forward(self, x, y):
-                res = torch.ops.quantized.add(self.quant1(x), self.quant2(y), 1.0, 0)
-                return self.dequant(res)
-
-        x = np.random.random(2).astype("float32")
-        y = np.random.random(2).astype("float32")
-        self.generic_test(QAddModule(), (x, y), input_names=["x", "y"])
-
-    def test_quantized_relu(self):
-        self.generic_unary_test(torch.nn.ReLU())
-
-    def test_qlinear_model(self):
-        class LinearModel(torch.nn.Module):
-            def __init__(self):
-                super(LinearModel, self).__init__()
-                self.qconfig = torch.quantization.default_qconfig
-                self.fc1 = torch.quantization.QuantWrapper(torch.nn.Linear(5, 10).to(dtype=torch.float))
-
-            def forward(self, x):
-                x = self.fc1(x)
-                return x
-
-        torch.backends.quantized.engine = "qnnpack"
-        qconfig = torch.quantization.default_qconfig
-        model = LinearModel()
-        model.qconfig = qconfig
-        model = torch.quantization.prepare(model)
-        model = torch.quantization.convert(model)
-
-        x_numpy = np.random.rand(1, 2, 5).astype(np.float32)
-        x = torch.from_numpy(x_numpy).to(dtype=torch.float)
-        outputs = model(x)
-        traced = torch.jit.trace(model, x)
-        buf = io.BytesIO()
-        torch.jit.save(traced, buf)
-        buf.seek(0)
-
-        model = torch.jit.load(buf)
-        input_names = ["x"]
-        f = io.BytesIO()
-        torch.onnx.export(model, x, f, input_names=input_names, example_outputs=outputs, operator_export_type=torch.onnx.OperatorExportTypes.ONNX_ATEN_FALLBACK)
-        f.seek(0)
-        onnx_model = onnx.load(f)
-        caffe_res = c2.run_model(onnx_model, dict(zip(input_names, x_numpy)))[0]
-        np.testing.assert_almost_equal(np.squeeze(outputs.numpy()), caffe_res, decimal=3)
 
 
 if __name__ == '__main__':
