+++ test/onnx/test_pytorch_onnx_caffe2_quantized.py
@@ -1,0 +1,131 @@
+from __future__ import print_function
+
+import numpy as np
+import unittest
+import torch.onnx
+import io
+
+import onnx
+import caffe2.python.onnx.backend as c2
+
+class TestQuantizedOps(unittest.TestCase):
+    def generic_test(self, model, sample_inputs, input_names=None):
+        pt_inputs = tuple(torch.from_numpy(x) for x in sample_inputs)
+        model.qconfig = torch.quantization.default_qconfig
+        q_model = torch.quantization.prepare(model, inplace=False)
+        q_model = torch.quantization.convert(q_model, inplace=False)
+        pytorch_res = q_model(*pt_inputs)
+        f = io.BytesIO()
+        torch.onnx.export(q_model, pt_inputs, f, input_names=input_names, operator_export_type=torch.onnx.OperatorExportTypes.ONNX_ATEN_FALLBACK)
+        f.seek(0)
+        onnx_model = onnx.load(f)
+        caffe_res = c2.run_model(onnx_model, dict(zip(input_names, sample_inputs)))[0]
+        np.testing.assert_almost_equal(pytorch_res.numpy(), caffe_res, decimal=3)
+
+    def generic_unary_test(self, op):
+        class QModule(torch.nn.Module):
+            def __init__(self, op):
+                super(QModule, self).__init__()
+                self.quant1 = torch.quantization.QuantStub()
+                self.op = op
+                self.dequant = torch.quantization.DeQuantStub()
+
+            def forward(self, x):
+                res = self.op(self.quant1(x))
+                return self.dequant(res)
+
+        x = np.random.random((1, 2)).astype("float32")
+        self.generic_test(QModule(op), (x,), input_names=["x"])
+
+    def test_quantized_add(self):
+        class QAddModule(torch.nn.Module):
+            def __init__(self):
+                super(QAddModule, self).__init__()
+                self.quant1 = torch.quantization.QuantStub()
+                self.quant2 = torch.quantization.QuantStub()
+                self.dequant = torch.quantization.DeQuantStub()
+
+            def forward(self, x, y):
+                res = torch.ops.quantized.add(self.quant1(x), self.quant2(y), 1.0, 0)
+                return self.dequant(res)
+
+        x = np.random.random(2).astype("float32")
+        y = np.random.random(2).astype("float32")
+        self.generic_test(QAddModule(), (x, y), input_names=["x", "y"])
+
+    def test_quantized_relu(self):
+        self.generic_unary_test(torch.nn.ReLU())
+
+    def test_qlinear_model(self):
+        class LinearModel(torch.nn.Module):
+            def __init__(self):
+                super(LinearModel, self).__init__()
+                self.qconfig = torch.quantization.default_qconfig
+                self.fc1 = torch.quantization.QuantWrapper(torch.nn.Linear(5, 10).to(dtype=torch.float))
+
+            def forward(self, x):
+                x = self.fc1(x)
+                return x
+
+        torch.backends.quantized.engine = "qnnpack"
+        qconfig = torch.quantization.default_qconfig
+        model = LinearModel()
+        model.qconfig = qconfig
+        model = torch.quantization.prepare(model)
+        model = torch.quantization.convert(model)
+
+        x_numpy = np.random.rand(1, 2, 5).astype(np.float32)
+        x = torch.from_numpy(x_numpy).to(dtype=torch.float)
+        outputs = model(x)
+        traced = torch.jit.trace(model, x)
+        buf = io.BytesIO()
+        torch.jit.save(traced, buf)
+        buf.seek(0)
+
+        model = torch.jit.load(buf)
+        input_names = ["x"]
+        f = io.BytesIO()
+        torch.onnx.export(model, x, f, input_names=input_names, example_outputs=outputs, operator_export_type=torch.onnx.OperatorExportTypes.ONNX_ATEN_FALLBACK)
+        f.seek(0)
+        onnx_model = onnx.load(f)
+        caffe_res = c2.run_model(onnx_model, dict(zip(input_names, x_numpy)))[0]
+        np.testing.assert_almost_equal(np.squeeze(outputs.numpy()), caffe_res, decimal=3)
+
+    def test_qconv_model(self):
+        class ConvModel(torch.nn.Module):
+            def __init__(self):
+                super(ConvModel, self).__init__()
+                self.qconfig = torch.quantization.default_qconfig
+                self.fc1 = torch.quantization.QuantWrapper(torch.nn.Conv2d(3, 5, 2, bias=False).to(dtype=torch.float))
+
+            def forward(self, x):
+                x = self.fc1(x)
+                return x
+        torch.backends.quantized.engine = "qnnpack"
+        qconfig = torch.quantization.default_qconfig
+        model = ConvModel()
+        model.qconfig = qconfig
+        model = torch.quantization.prepare(model)
+        model = torch.quantization.convert(model)
+
+        x_numpy = np.random.rand(1, 3, 6, 6).astype(np.float32)
+
+        x = torch.from_numpy(x_numpy).to(dtype=torch.float)
+        outputs = model(x)
+        traced = torch.jit.trace(model, x)
+        buf = io.BytesIO()
+        torch.jit.save(traced, buf)
+        buf.seek(0)
+        model = torch.jit.load(buf)
+        input_names = ["x"]
+        f = io.BytesIO()
+        torch.onnx.export(model, x, f, input_names=input_names, example_outputs=outputs, operator_export_type=torch.onnx.OperatorExportTypes.ONNX_ATEN_FALLBACK)
+        f.seek(0)
+        onnx_model = onnx.load(f)
+        x_c2 = np.transpose(x_numpy, [0, 2, 3, 1])
+        y = np.expand_dims(x_c2, axis=0)
+        caffe_res = c2.run_model(onnx_model, dict(zip(input_names, y)))[0]
+        np.testing.assert_almost_equal(outputs.permute(0, 2, 3, 1).numpy(), caffe_res, decimal=3)
+
+if __name__ == '__main__':
+    unittest.main()
